Improve network handling in API wrapper (#10)
This replaces a number of force-unwraps and vague error-handling paths
in the API wrapper with a much more structured `Result` type. While the
error case handling is far from exhaustive, the general structure should
be far more robust both in the event of connectivity issues and
unexpected backend responses.

Fixes #4, and makes some progress on #9.