Call delegate failure in remaining cases (#13)
This should handle all of the remaining cases where the delegate method
might not be called in a failure scenario (excepting the case where
`ASAuthorization` doesn't call _our_ delegate method, which is an
upstream bug). The error codes are not correct in all situations, but
theoretically will always result in a complete cycle.