Convert delegate API to async/await (#17)
This was the intended API from the start, and Continuations ended up
being the trick to bridge the delegate-based ASAuthorization APIs into
the SnapAuth SDKs.

Now, instead of setting the delegate and calling the method, the client
code looks more like `let authInfo = await
snapAuth.startauth(.handle(username))`. It should be more
error-resistant (you can't fail to set the delegate, and will get a
compiler warning if you ignore the result) and a bit easier and more
streamlined to implement.

The AutoFill APIs remain delegate-based (following a structural
adjustment), similar to how the JS/TS is passed a callback This is
definitely open to adjustment based on feedback.

Fixes #18.