Clean up last remaining unguarded force-unwraps (#21)
Force-unwraps can lead to runtime crashes if the underlying value is
`nil`. I went through and searched for every `!` in the codebase. Most
already had a guard of some kind (some a literal `guard`, others with
some other conditional). This adds some checks for the last ones:

- buildAuthRequests is not useful without allowCredentials in the
hardware key path, so that path is skipped if no allowCredential is
provided (the API will always provide them when a username or id is
sent, so this is more of a typechecking formality)
- the ASPresentationAnchor needs to be set for the UI to load, so an
additional fallback was put in place in case the internal value is nil


Fixes #9