Rework autofill to also use async/await (#31)
As noted in #29, the initial design of the autofill interactions used a
delegate-based system, rather than async/await like the modal APIs. This
was based on an incorrect assumption (carried over from web, where it
may also be flawed) on how promises resolve. Upon further
experimentation, it seems perfectly fine and non-problematic to have a
promise that might take minutes to resolve and not have it block other
interactions.

This (breaking) change adjusts the autofill API to also use async/await
instead of delegates. Client code is a lot more straightforward, and the
internals become more reliable since a bunch of conditional paths are
eliminated.

In the same vein, I've also added a new `.unsupportedOnPlatform` error
code so that clients can have fewer availability checks, part of #30.
The API as a whole is still `#if`'d to iOS+visionOS, but that will get
lifted eventually in a separate PR (see #16).