Pull error conversion out into an extension to simplify controller (#32)
We encapsulate the `ASAuthenticationError.Code` values into our own
`SnapAuthError` format. This change moves the conversion outside of the
AS delegate callback into an extension defined near our own error codes.
Doing so encourages a single source of truth in case future
functionality requires a similar conversion in a different path.

It also adds support for a newly-defined case in iOS 18 and same-year
equivalents.