Remove erroneous assertion (#34)
#31 adjusted the assertion during sending errors back, and reintroduced
the data race from canceling the autofill request due to starting a
modal one (explicitly returning a failure to the continuation, clearing
it out, and canceling the authController - which went back to sendError)

This removes the assertion, since it was logically unsound, not to
mention unhelpful. I've done another bunch of manual testing in a sample
app and can no longer reproduce the issue. I'm not sure how to cover
this in automated tests to prevent another regression, but I'll file a
ticket in case someone finds a way.